post rebase

+++ src/Types.hs
@@ -3,7 +3,7 @@ module Types where
 import Data.Matrix
 import qualified Data.Vector as V
 
-type Temp        = Float
+type Temp        = Double
 type Time        = Int
 type Graph    = Matrix Int
 type Coloring = V.Vector Int
